+++ tests/Clusterers/DBSCANTest.php
@@ -17,6 +17,7 @@ use Rubix\ML\Datasets\Generators\Circle;
 use Rubix\ML\Datasets\Generators\Agglomerate;
 use Rubix\ML\CrossValidation\Metrics\VMeasure;
 use Rubix\ML\Exceptions\InvalidArgumentException;
+use Rubix\ML\Exceptions\RuntimeException;
 use PHPUnit\Framework\TestCase;
 
 #[Group('Clusterers')]
@@ -24,14 +25,19 @@ use PHPUnit\Framework\TestCase;
 class DBSCANTest extends TestCase
 {
     /**
+     * The number of samples in the training set.
+     */
+    protected const int TRAIN_SIZE = 512;
+
+    /**
      * The number of samples in the validation set.
      */
-    protected const int TEST_SIZE = 512;
+    protected const int TEST_SIZE = 256;
 
     /**
      * The minimum validation score required to pass the test.
      */
-    protected const float MIN_SCORE = 0.65;
+    protected const float MIN_SCORE = 0.9;
 
     /**
      * Constant used to see the random number generator.
@@ -46,7 +52,7 @@ class DBSCANTest extends TestCase
 
     protected function setUp() : void
     {
-        generators: $this->generator = new Agglomerate(
+        $this->generator = new Agglomerate(
             [
                 'inner' => new Circle(x: 0.0, y: 0.0, scale: 1.0, noise: 0.01),
                 'middle' => new Circle(x: 0.0, y: 0.0, scale: 5.0, noise: 0.05),
@@ -59,6 +65,12 @@ class DBSCANTest extends TestCase
         $this->metric = new VMeasure();
 
         srand(self::RANDOM_SEED);
+    }
+
+    #[Test]
+    public function preConditions() : void
+    {
+        $this->assertFalse($this->estimator->trained());
     }
 
     #[Test]
@@ -98,9 +110,14 @@ class DBSCANTest extends TestCase
     }
 
     #[Test]
-    public function predict() : void
+    public function trainPredict() : void
     {
+        $training = $this->generator->generate(self::TRAIN_SIZE);
         $testing = $this->generator->generate(self::TEST_SIZE);
+
+        $this->estimator->train($training);
+
+        $this->assertTrue($this->estimator->trained());
 
         $predictions = $this->estimator->predict($testing);
 
@@ -113,11 +130,49 @@ class DBSCANTest extends TestCase
     }
 
     #[Test]
+    public function trainIncompatible() : void
+    {
+        $this->expectException(InvalidArgumentException::class);
+
+        $this->estimator->train(Unlabeled::quick(samples: [['bad']]));
+    }
+
+    #[Test]
+    public function predictUntrained() : void
+    {
+        $this->expectException(RuntimeException::class);
+
+        $this->estimator->predict(Unlabeled::quick(samples: [[1.0, 2.0]]));
+    }
+
+    #[Test]
     #[TestDox('Throws an exception when predicting with incompatible data')]
     public function predictIncompatible() : void
     {
         $this->expectException(InvalidArgumentException::class);
 
-        $this->estimator->predict(Unlabeled::quick(samples: [['bad']]));
+        $training = $this->generator->generate(self::TRAIN_SIZE);
+
+        $this->estimator->train($training);
+
+        $this->estimator->predict(Unlabeled::quick(samples: [[1.0]]));
+    }
+
+    #[Test]
+    public function restoreStateFromSerializedModel() : void
+    {
+        $training = $this->generator->generate(self::TRAIN_SIZE);
+
+        $this->estimator->train($training);
+
+        $this->assertTrue($this->estimator->trained());
+
+        $restored = unserialize(serialize($this->estimator));
+
+        $this->assertTrue($restored->trained());
+
+        $testing = $this->generator->generate(self::TEST_SIZE);
+
+        $this->assertEquals($this->estimator->predict($testing), $restored->predict($testing));
     }
 }