+++ CHANGELOG.md
@@ -58,6 +58,7 @@
     - Iterative Learners renamed steps() method to progress()
     - Changed default gradient-based `minChange` from 1e-4 to 1e-5
     - `Prior` is now default Strategy of Missing Data Imputer
+    - `DBSCAN` is now a `Learner` and `Persistable`, `predict()` requires prior `train()`
     
 - 2.5.15
     - Optimize random weighted subset with replacement